rbd: lay out header probe infrastructure
This defines a new function rbd_dev_header_probe() as a top-level
function for populating an rbd header.  It first checks for the
existence of a format 2 rbd image id object.  If it exists, the
image is assumed to be a format 2 rbd image, and another function
rbd_dev_header_v2() is called to finish populating header data for
that image.  If it does not exist, it is assumed to be an old
(format 1) rbd image, calls a similar function rbd_dev_header_v1()
to populate its header information.

So far, the format 2 images are not really supported; this is
laying the infrastructure for fleshing out that support.

Signed-off-by: Alex Elder <elder@inktank.com>