libceph: close socket directly from ceph_con_close()
When we are directed to close a connect, close the socket directly while
holding the con mutex.  This is simpler and less fragile than the old
method of setting the flag and queueing con work.  In particular, we will
frequently re-open the same connection and need to close+open the socket
in con_work(), which doesn't work well.  We thus prevent a situation where
we encounter the SOCK_CLOSED flag and OPENING state (which isn't currently
handled).

We also carry the mutex for the entire duration of ceph_con_open(),
including the period where we are twiddling state bits.  This avoids the
possibility of another thread seeing an undefined state, moving the
con->state manipulation under the protectino of the mutex (in this case;
there are still other places where it is not protected).

This also appears to fix not-completely-explained crashes with socket
failure injection on the server side.

Signed-off-by: Sage Weil <sage@inktank.com>